move external capabilities to ai_orchestrator.go, remove duplicate get price function
